pyproject.toml
[tool.poetry]
name = "agent-workforce"
version = "0.1.0"
description = "A platform for managing and executing AI agent crews"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
pydantic = "^2.4.2"
crewai = "^0.119.0"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.6"
sqlalchemy = "^2.0.23"
asyncpg = "^0.29.0"
alembic = "^1.12.1"
aiosqlite = "^0.20.0"
importlib-metadata = ">=6.0,<8.7.0"
setuptools = "^80.3.0"
greenlet = "^3.2.1"
langchain-anthropic = "^0.3.13"
langchain-openai = "^0.3.16"
duckduckgo-search = "^4.4.3"
pandas = "^2.1.3"
atlassian-python-api = "^4.0.3"
langchain-core = "^0.3.59"
langchain = "^0.3.25"
langchain-community = "^0.3.23"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3 